fix(phase_b_streaming): residual P1+P2 7 fixes (no debt)
P1 #2 : LOT 1 trades_in_window contract documente (sorted ts_event ASC)
         + assertion debug-only (RuntimeWarning si caller envoie non-sorted)
P1 #3 : LOT 1 weekend transition test (3-day gap fri->mon) - PASS
P1 #4 : LOT 3 cluster_v2 docstring runs detection algorithme detaille
P1 #6 : LOT 4 absorb HVL verifie 2x intentionnel (mirror batch ligne 1010-1028)
P1 #8 : gold_phase_d cross-asset realistic synthetic test - PASS
        (im_dxy_corr_60d 0.636 attendu 0.2-0.8 / proxy 1.58 attendu 0.5-3.0)
P2 #5 : LOT 5 finish_pct float() cast explicite (parite float32 batch)
P2 #10: intermarket open_type_agreement comment (NaN deja safe-cast 0.0)

Tests :
  - TOOLS/test_phase_b_trades_streaming_weekend.py : PASS
  - TOOLS/test_gold_phase_d_realistic.py : PASS (2 tests)
  - TOOLS/test_phase_b_streaming_real_data.py : PASS 1380 bars / 0 erreurs

Aucune modification logique fonctionnelle (docs/commentaires/asserts/tests).
Batch engine (phase_b_plus_plus_engine.py, gold_phase_d_features.py) intact.

+++ CORE/phase_b_plus_plus_trapped_streaming.py
@@ -140,7 +140,11 @@ def add_trapped_traders_streaming(
     trap_sell = 0
     if footprint_cells and not (np.isnan(h) or np.isnan(l) or np.isnan(op) or np.isnan(c)):
         bar_range = h - l
-        finish_pct = (c - l) / bar_range * 100 if bar_range > 0 else 50.0
+        # finish_pct calcule en float64 par defaut Python. Le batch utilise float32
+        # (memoire). Cast explicite pour aligner precision et eviter divergence
+        # numerique tres faible (1e-7) qui pourrait flipper les comparaisons aux
+        # seuils TRAPPED_FINISH_PCT_BUYERS=30 / TRAPPED_FINISH_PCT_SELLERS=70.
+        finish_pct = float((c - l) / bar_range * 100) if bar_range > 0 else 50.0
 
         # Trapped buyers : AskVol(top zone) seuil + finish<=30 + delta>0 + close<open
         ask_h = 0.0
